==> Code/13 code.ipynb <==
# Classification Implementation

import numpy as np
import mne
from scipy import signal
from scipy.interpolate import RectBivariateSpline
from mne.filter import resample, filter_data
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt
from lspopt import spectrogram_lspopt
from matplotlib.colors import Normalize, ListedColormap

import logging
LOGGING_TYPES = dict(DEBUG=logging.DEBUG, INFO=logging.INFO, WARNING=logging.WARNING,
                     ERROR=logging.ERROR, CRITICAL=logging.CRITICAL)
logger = logging.getLogger('yasa')

%matplotlib qt


# load reference_df     
reference_df = pd.read_csv("reference_df.csv", index_col="name")
reference_df.head(3)

# load csv    
rankings_df = pd.read_csv("rankings_df.csv", index_col="method_name")
rankings_df.head(3)

# to see id's
idx = reference_df.index.to_list()

# to load hypno:
hypno_loc = reference_df.loc[idx[10], "hypno"]
hypno_30s = np.loadtxt(hypno_loc, dtype ='int')[:, 0]

# # to load features:
df_feat_loc = reference_df.loc[idx[10], "df_feat"]
df_feat = pd.read_csv(df_feat_loc, index_col=False)


hypno_30s.shape

df_feat.shape

Omitting artifcat epochs

df_feat = df_feat.loc[hypno_30s!=-1]
hypno_30s = hypno_30s[hypno_30s!=-1]

print('Class labels:', np.unique(hypno_30s))

from sklearn.model_selection import train_test_split

X_train, X_test, y_train, y_test = train_test_split(
    df_feat, hypno_30s, test_size=0.2, random_state=1, stratify=hypno_30s
)


print("Labels counts in y:", np.bincount(hypno_30s))
print("Labels counts in y_train:", np.bincount(y_train))
print("Labels counts in y_test:", np.bincount(y_test))

from sklearn.preprocessing import StandardScaler

sc = StandardScaler()
sc.fit(df_feat)
X_train_std = sc.transform(X_train)
X_test_std = sc.transform(X_test)

# Perceptron

from sklearn.linear_model import Perceptron

ppn = Perceptron(eta0=0.1, random_state=1)
ppn.fit(X_train_std, y_train)

y_pred = ppn.predict(X_test_std)
print('Misclassified examples: %d' % (y_test != y_pred).sum())

from sklearn.metrics import accuracy_score

print('Accuracy: %.3f' % accuracy_score(y_test, y_pred))

print('Accuracy: %.3f' % ppn.score(X_test_std, y_test))

from sklearn.metrics import precision_score, recall_score, f1_score
from sklearn.metrics import matthews_corrcoef

pre_val = precision_score(y_true=y_test, y_pred=y_pred, average="micro")
print(f'Precision: {pre_val:.3f}')

rec_val = recall_score(y_true=y_test, y_pred=y_pred, average="micro")
print(f'Recall: {rec_val:.3f}')

f1_val = f1_score(y_true=y_test, y_pred=y_pred, average="micro")
print(f'F1: {f1_val:.3f}')

mcc_val = matthews_corrcoef(y_true=y_test, y_pred=y_pred)
print(f'MCC: {mcc_val:.3f}')

# Logistic Regression

from sklearn.linear_model import LogisticRegression

lr = LogisticRegression()  # C=100.0, solver='lbfgs', multi_class='ovr')
lr.fit(X_train_std, y_train)

print(lr.predict_proba(X_test_std[:3, :]).argmax(axis=1))


y_pred = lr.predict(X_test_std)
print("Misclassified examples: %d" % (y_test != y_pred).sum())

print("Accuracy: %.3f" % accuracy_score(y_test, y_pred))

pre_val = precision_score(y_true=y_test, y_pred=y_pred, average="micro")
print(f'Precision: {pre_val:.3f}')

rec_val = recall_score(y_true=y_test, y_pred=y_pred, average="micro")
print(f'Recall: {rec_val:.3f}')

f1_val = f1_score(y_true=y_test, y_pred=y_pred, average="micro")
print(f'F1: {f1_val:.3f}')

mcc_val = matthews_corrcoef(y_true=y_test, y_pred=y_pred)
print(f'MCC: {mcc_val:.3f}')


weights, params = [], []
for c in np.arange(-5, 5):
    lr = LogisticRegression(C=10.0**c, multi_class="ovr")
    lr.fit(X_train_std, y_train)
    weights.append(lr.coef_[1])
    params.append(10.0**c)

weights = np.array(weights)
params = np.array(params)
plt.figure()
plt.plot(params, weights[:, 0], label="std")
plt.plot(params, weights[:, 1], linestyle="--", label="mean")
plt.ylabel("Weight coefficient")
plt.xlabel("C")
plt.legend(loc="upper left")
plt.xscale("log")
# plt.savefig('figures/03_08.png', dpi=300)
plt.show()


# Maximum margin classification with support vector machines

from sklearn.svm import SVC

svm = SVC(kernel='linear', C=1.0, random_state=1)
svm.fit(X_train_std, y_train)

y_pred = lr.predict(X_test_std)
print('Misclassified examples: %d' % (y_test != y_pred).sum())

print('Accuracy: %.3f' % accuracy_score(y_test, y_pred))

pre_val = precision_score(y_true=y_test, y_pred=y_pred, average="micro")
print(f'Precision: {pre_val:.3f}')

rec_val = recall_score(y_true=y_test, y_pred=y_pred, average="micro")
print(f'Recall: {rec_val:.3f}')

f1_val = f1_score(y_true=y_test, y_pred=y_pred, average="micro")
print(f'F1: {f1_val:.3f}')

mcc_val = matthews_corrcoef(y_true=y_test, y_pred=y_pred)
print(f'MCC: {mcc_val:.3f}')

from sklearn.linear_model import SGDClassifier

ppn = SGDClassifier(loss='perceptron')
ppn.fit(X_train_std, y_train)
y_pred = ppn.predict(X_test_std)
print('ppn Misclassified examples: %d' % (y_test != y_pred).sum())
print('ppn Accuracy: %.3f' % accuracy_score(y_test, y_pred))

lr = SGDClassifier(loss='log')
lr.fit(X_train_std, y_train)
y_pred = lr.predict(X_test_std)
print('lr Misclassified examples: %d' % (y_test != y_pred).sum())
print('lr Accuracy: %.3f' % accuracy_score(y_test, y_pred))

svm = SGDClassifier(loss='hinge')
svm.fit(X_train_std, y_train)
y_pred = svm.predict(X_test_std)
print('svm Misclassified examples: %d' % (y_test != y_pred).sum())
print('svm Accuracy: %.3f' % accuracy_score(y_test, y_pred))


# Plotting feature space 

feat1 = "sb"
feat2 = "alpha"

df_feat1 = df_feat.copy()
df_feat1 = df_feat1[[feat1, feat2]]

sc1 = StandardScaler()
sc1.fit(df_feat1)
X = sc1.transform(df_feat1)
y = hypno_30s

plt.figure(figsize=(7,5))
plt.scatter(X[y == 0, 0], X[y == 0, 1], marker="s", label="Class 0")
plt.scatter(X[y == 1, 0], X[y == 1, 1], marker="o", label="Class 1")
plt.scatter(X[y == 2, 0], X[y == 2, 1], marker="o", label="Class 2")
plt.scatter(X[y == 3, 0], X[y == 3, 1], marker="o", label="Class 3")
plt.scatter(X[y == 4, 0], X[y == 4, 1], marker="o", label="Class 4")

plt.xlabel(f"Feature 1 - {feat1}")
plt.ylabel(f"Feature 2 - {feat2}")
plt.legend(loc="best")
plt.tight_layout()
# plt.savefig(f"{feat1}_vs_{feat2} features.png", dpi=300)
# plt.savefig(f"{feat1}_vs_{feat2} features.svg")
plt.show()


fig, axes = plt.subplots(5, 5, figsize=(10, 10))
for i, ax in enumerate(axes.ravel()):
    feat1 = rankings_df.columns.values.tolist()[int(i // 5 + 7)]
    feat2 = rankings_df.columns.values.tolist()[int(np.random.randint(1, 65))]

    df_feat1 = df_feat.copy()
    df_feat1 = df_feat1[[feat1, feat2]]

    sc1 = StandardScaler()
    sc1.fit(df_feat1)
    X = sc1.transform(df_feat1)
    y = hypno_30s

    ax.scatter(X[y == 0, 0], X[y == 0, 1], marker="o", label="Class 0")
    ax.scatter(X[y == 1, 0], X[y == 1, 1], marker="v", label="Class 1")
    ax.scatter(X[y == 2, 0], X[y == 2, 1], marker="s", label="Class 2")
    ax.scatter(X[y == 3, 0], X[y == 3, 1], marker="*", label="Class 3")
    ax.scatter(X[y == 4, 0], X[y == 4, 1], marker="p", label="Class 4")

    ax.set_yticks([])
    ax.set_xticks([])

    ax.set_xlabel(f"{feat1}")
    ax.set_ylabel(f"{feat2}")

handles, labels = ax.get_legend_handles_labels()
fig.legend(handles, labels, loc='lower right')

plt.tight_layout()
# plt.savefig(f"2D features space.png", dpi=300)
# plt.savefig(f"2D features space.svg")
plt.show()


# def plot_decision_regions(X, y, classifier, test_idx=None, resolution=0.02):

#     # setup marker generator and color map
#     markers = ('o', 's', '^', 'v', '<')
#     colors = ('red', 'blue', 'lightgreen', 'gray', 'cyan')
#     cmap = ListedColormap(colors[:len(np.unique(y))])

#     # plot the decision surface
#     x1_min, x1_max = X[:, 0].min() - 1, X[:, 0].max() + 1
#     x2_min, x2_max = X[:, 1].min() - 1, X[:, 1].max() + 1
#     xx1, xx2 = np.meshgrid(np.arange(x1_min, x1_max, resolution),
#                            np.arange(x2_min, x2_max, resolution))
#     lab = classifier.predict(np.array([xx1.ravel(), xx2.ravel()]).T)
#     lab = lab.reshape(xx1.shape)
#     plt.contourf(xx1, xx2, lab, alpha=0.3, cmap=cmap)
#     plt.xlim(xx1.min(), xx1.max())
#     plt.ylim(xx2.min(), xx2.max())

#     # plot class examples
#     for idx, cl in enumerate(np.unique(y)):
#         plt.scatter(x=X[y == cl, 0], 
#                     y=X[y == cl, 1],
#                     alpha=0.8, 
#                     c=colors[idx],
#                     marker=markers[idx], 
#                     label=f'Class {cl}', 
#                     edgecolor='black')

#     # highlight test examples
#     if test_idx:
#         # plot all examples
#         X_test, y_test = X[test_idx, :], y[test_idx]

#         plt.scatter(X_test[:, 0],
#                     X_test[:, 1],
#                     c='none',
#                     edgecolor='black',
#                     alpha=1.0,
#                     linewidth=1,
#                     marker='o',
#                     s=100, 
#                     label='Test set')     

# svm = SVC(kernel='rbf', random_state=1, gamma=0.10, C=10.0)
# svm.fit(X, y)
# plot_decision_regions(X, y,
#                       classifier=svm)

# plt.legend(loc='upper left')
# plt.tight_layout()
# #plt.savefig('figures/03_14.png', dpi=300)
# plt.show()   

# Solving non-linear problems using a kernel SVM

# We select 20 first columns based on the statistical significance in classification 
columns = rankings_df.columns[:20]

# let’s divide the dataset into a separate training dataset (80 percent of the data) and a separate test dataset (20 percent of the data)
X_train, X_test, y_train, y_test = train_test_split(
    df_feat[columns], hypno_30s, test_size=0.2, random_state=1, stratify=hypno_30s
)
 
X_train, X_test = X_train[columns], X_test[columns]

# we will standardize the columns in dataset before we feed them to a classifier
sc = StandardScaler()
sc.fit(df_feat[columns])
X_train_std = sc.transform(X_train)
X_test_std = sc.transform(X_test)


X_train.head(3)

from sklearn.svm import SVC

svm = SVC(kernel='rbf', random_state=1, gamma=0.2, C=1.0)
svm.fit(X_train_std, y_train)
y_pred = svm.predict(X_test_std)
print('Misclassified examples: %d' % (y_test != y_pred).sum())
print('Accuracy: %.3f' % accuracy_score(y_test, y_pred))


## Building a decision tree

from sklearn.tree import DecisionTreeClassifier

tree_model = DecisionTreeClassifier(criterion='gini', 
                                    max_depth=5, 
                                    random_state=1)
tree_model.fit(X_train, y_train)
y_pred = tree_model.predict(X_test)
print('Misclassified examples: %d' % (y_test != y_pred).sum())
print('Accuracy: %.3f' % accuracy_score(y_test, y_pred))

from sklearn import tree

tree.plot_tree(tree_model, filled=True)

# plt.savefig('figures/03_21_1.pdf')
plt.show()


## Combining weak to strong learners via random forests

from sklearn.ensemble import RandomForestClassifier

forest = RandomForestClassifier(n_estimators=25, 
                                random_state=1,
                                n_jobs=2)
forest.fit(X_train, y_train)
y_pred = forest.predict(X_test)
print('Misclassified examples: %d' % (y_test != y_pred).sum())
print('Accuracy: %.3f' % accuracy_score(y_test, y_pred))

# K-nearest neighbors - a lazy learning algorithm

from sklearn.neighbors import KNeighborsClassifier

knn = KNeighborsClassifier(n_neighbors=5, 
                           p=2, 
                           metric='minkowski')
knn.fit(X_train_std, y_train)
y_pred = knn.predict(X_test_std)
print('Misclassified examples: %d' % (y_test != y_pred).sum())
print('Accuracy: %.3f' % accuracy_score(y_test, y_pred))

# Assessing feature importance with Random Forests

from sklearn.ensemble import RandomForestClassifier

columns = rankings_df.columns[:50]

X_train, X_test, y_train, y_test = train_test_split(
    df_feat[columns], hypno_30s, test_size=0.2, random_state=1, stratify=hypno_30s
)

X_train, X_test = X_train[columns], X_test[columns]

feat_labels = df_feat[columns].columns[:]

forest = RandomForestClassifier(n_estimators=500, random_state=1)

forest.fit(X_train, y_train)
importances = forest.feature_importances_

indices = np.argsort(importances)[::-1]

for f in range(X_train.shape[1]):
    print(
        "%2d) %-*s %f" % (f + 1, 30, feat_labels[indices[f]], importances[indices[f]])
    )

plt.title("Feature importance")
plt.bar(range(X_train.shape[1]), importances[indices], align="center")

plt.xticks(range(X_train.shape[1]), feat_labels[indices], rotation=90)
plt.xlim([-1, X_train.shape[1]])
plt.tight_layout()
# plt.savefig('figures/04_10.png', dpi=300)
plt.show()


from sklearn.feature_selection import SelectFromModel

sfm = SelectFromModel(forest, threshold=0.02, prefit=True)
X_selected = sfm.transform(X_train)
print('Number of features that meet this threshold criterion:', 
      X_selected.shape[1])

for f in range(X_selected.shape[1]):
    print("%2d) %-*s %f" % (f + 1, 30, 
                            feat_labels[indices[f]], 
                            importances[indices[f]]))
