==> Code/22 code.ipynb <==
# After feature sorting (code 17), see the correlation between top features. Then ommit features with high correlation.

Also elinimating the highly correlated ones.

import numpy as np
import mne
from scipy import signal
from scipy.interpolate import RectBivariateSpline
from mne.filter import resample, filter_data
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt
from lspopt import spectrogram_lspopt
from matplotlib.colors import Normalize, ListedColormap

import logging
LOGGING_TYPES = dict(DEBUG=logging.DEBUG, INFO=logging.INFO, WARNING=logging.WARNING,
                     ERROR=logging.ERROR, CRITICAL=logging.CRITICAL)
logger = logging.getLogger('yasa')

%matplotlib qt


# load reference_df     
reference_df = pd.read_csv("reference_df.csv", index_col="name")
reference_df.head(3)

# load csv    
rankings_df = pd.read_csv("rankings_df aug.csv", index_col="method_name")
rankings_df.head(3)

rankings_df.columns

# load features

### to load features for augmented eeg:
df_feat_loc = reference_df.iloc[0].df_feat
df_feat_loc = df_feat_loc.split(".")[0] + " aug.csv"
df_feat = pd.read_csv(df_feat_loc, index_col=False)

# Correlation Analysis of Top features

def pearson_corr(df, output_var="hmob", thresh=0.8):
    # Using Pearson Correlation
    corr_matrix = df.corr()

    # Correlation with output variable
    cor_target = abs(corr_matrix[output_var])
    # Selecting highly correlated features
    relevant_features = cor_target[cor_target >= thresh]
    print(f"RELEVANT ONES to {output_var}:")
    print(relevant_features)
    print()

    return relevant_features


def pearson_correlation_plot(df, thresh=0.95, fname="correlation_plot"):
    plt.figure(figsize=(12, 10))
    plt.suptitle(f"Pearson Correlation Plot for Top Uncorrelated Features 90%")
    corr_matrix = df.corr()
    corr_matrix[abs(corr_matrix) >= thresh] = abs(corr_matrix)
    corr_matrix[abs(corr_matrix) < thresh] = 0
    # print(corr_matrix.shape)
    g = sns.heatmap(
        abs(corr_matrix), cmap=plt.cm.Reds, xticklabels=True, yticklabels=True
    )
    g.set_xticklabels(g.get_xmajorticklabels(), fontsize=7)
    g.set_yticklabels(g.get_ymajorticklabels(), fontsize=7)
    plt.tight_layout()
    plt.savefig(f"{fname}.png", format="png")
    plt.savefig(f"{fname}.svg", format="svg")
    plt.show()


# pearson_correlation_plot(df_feat.loc[:, top_features], thresh=0.7)


top_features = rankings_df.columns.values.tolist()
for idx, feature in enumerate(top_features): 
    # Sort df_feat based on top-features in an ascending order
    df_feat = df_feat.loc[:, top_features] 
    # Find highly correlated features from current onward
    highly_corr_feats = pearson_corr(df_feat.loc[:, feature:], output_var=feature, thresh=0.90)
    # Omit ccurrent (itself) and convert to list
    highly_corr_feats = highly_corr_feats.index.tolist()[1:] # convert to list, omit itself
    # Remove correlated features and update `top_features` list
    for feat2remove in highly_corr_feats:
        top_features.remove(feat2remove)
    

pearson_correlation_plot(df_feat.loc[:, top_features], thresh=0, fname="correlation_plot thresh {thresh} top uncorr 90")

top_feat_initial = rankings_df.columns.values.tolist()
top_feat_updated = top_features

for i, feat in enumerate(top_feat_initial):
    if feat in top_feat_updated:
        print(f"{i}- {feat}: ✅")
    else:
        print(f"{i}- {feat}: ❌")


# write top features into a file
with open('top_features correlation_90.txt', 'w') as f:
    for s in top_feat_updated:
        f.write(s + '\n')

# read top features from file
with open('top_features correlation_90.txt', 'r') as f:
    top_feat = [line.rstrip('\n') for line in f]


len(top_feat)
